# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Enterprise Candidate Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type CandidateStatus = typeof CandidateStatus[keyof typeof CandidateStatus];


export const CandidateStatus = {
  applied: 'applied',
  screening: 'screening',
  interview: 'interview',
  offer: 'offer',
  selected: 'selected',
  rejected: 'rejected',
} as const;

export interface Candidate {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  skills?: string[];
  /** @nullable */
  yearsOfExperience?: number | null;
  /** @nullable */
  currentCompany?: string | null;
  /** @nullable */
  resumeUrl?: string | null;
  /** @nullable */
  position?: string | null;
  status: CandidateStatus;
  createdAt: string;
  updatedAt: string;
}

export interface Note {
  id: number;
  candidateId: number;
  content: string;
  authorName: string;
  createdAt: string;
  updatedAt: string;
}

export interface ActivityLog {
  id: number;
  candidateId: number;
  candidateName: string;
  action: string;
  /** @nullable */
  oldValue?: string | null;
  /** @nullable */
  newValue?: string | null;
  createdAt: string;
}

export interface CandidateDetail {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  skills?: string[];
  /** @nullable */
  yearsOfExperience?: number | null;
  /** @nullable */
  currentCompany?: string | null;
  /** @nullable */
  resumeUrl?: string | null;
  /** @nullable */
  position?: string | null;
  status: CandidateStatus;
  createdAt: string;
  updatedAt: string;
  notes: Note[];
  activityLogs: ActivityLog[];
}

export interface CandidateList {
  data: Candidate[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CandidateInput {
  /** @minLength 1 */
  name: string;
  email: string;
  phone?: string;
  skills?: string[];
  /** @minimum 0 */
  yearsOfExperience?: number;
  currentCompany?: string;
  resumeUrl?: string;
  position?: string;
  status?: CandidateStatus;
}

export interface CandidateUpdate {
  /** @minLength 1 */
  name?: string;
  email?: string;
  phone?: string;
  skills?: string[];
  /** @minimum 0 */
  yearsOfExperience?: number;
  currentCompany?: string;
  resumeUrl?: string;
  position?: string;
  status?: CandidateStatus;
}

export interface CandidateStatusUpdate {
  status: CandidateStatus;
}

export interface NoteInput {
  /** @minLength 1 */
  content: string;
  /** @minLength 1 */
  authorName: string;
}

export interface NoteUpdate {
  /** @minLength 1 */
  content: string;
}

export interface StatusCounts {
  applied: number;
  screening: number;
  interview: number;
  offer: number;
  selected: number;
  rejected: number;
}

export interface DashboardStats {
  totalCandidates: number;
  byStatus: StatusCounts;
  recentCandidates: Candidate[];
  recentActivity: ActivityLog[];
}

export interface KanbanColumn {
  status: CandidateStatus;
  candidates: Candidate[];
}

export interface KanbanBoard {
  columns: KanbanColumn[];
}

export type ListCandidatesParams = {
page?: number;
limit?: number;
search?: string;
status?: ListCandidatesStatus;
sortBy?: ListCandidatesSortBy;
sortOrder?: ListCandidatesSortOrder;
};

export type ListCandidatesStatus = typeof ListCandidatesStatus[keyof typeof ListCandidatesStatus];


export const ListCandidatesStatus = {
  applied: 'applied',
  screening: 'screening',
  interview: 'interview',
  offer: 'offer',
  selected: 'selected',
  rejected: 'rejected',
} as const;

export type ListCandidatesSortBy = typeof ListCandidatesSortBy[keyof typeof ListCandidatesSortBy];


export const ListCandidatesSortBy = {
  name: 'name',
  createdAt: 'createdAt',
  updatedAt: 'updatedAt',
  yearsOfExperience: 'yearsOfExperience',
} as const;

export type ListCandidatesSortOrder = typeof ListCandidatesSortOrder[keyof typeof ListCandidatesSortOrder];


export const ListCandidatesSortOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;

export type ListActivityParams = {
limit?: number;
candidateId?: number;
};