# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Enterprise Candidate Management System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: candidates
    description: Candidate management
  - name: notes
    description: Candidate notes
  - name: dashboard
    description: Dashboard statistics
  - name: activity
    description: Activity logs
  - name: kanban
    description: Kanban board

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List candidates
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: search
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema:
            type: string
            enum: [applied, screening, interview, offer, selected, rejected]
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [name, createdAt, updatedAt, yearsOfExperience]
            default: createdAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Paginated list of candidates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateList"

    post:
      operationId: createCandidate
      tags: [candidates]
      summary: Create a candidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateInput"
      responses:
        "201":
          description: Created candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /candidates/{id}:
    get:
      operationId: getCandidate
      tags: [candidates]
      summary: Get a candidate by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Candidate detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateCandidate
      tags: [candidates]
      summary: Update a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateUpdate"
      responses:
        "200":
          description: Updated candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteCandidate
      tags: [candidates]
      summary: Delete a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  /candidates/{id}/status:
    patch:
      operationId: updateCandidateStatus
      tags: [candidates]
      summary: Update candidate status (Kanban move)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateStatusUpdate"
      responses:
        "200":
          description: Updated candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /candidates/{candidateId}/notes:
    get:
      operationId: listCandidateNotes
      tags: [notes]
      summary: List notes for a candidate
      parameters:
        - name: candidateId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"

    post:
      operationId: createCandidateNote
      tags: [notes]
      summary: Create a note for a candidate
      parameters:
        - name: candidateId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"

  /notes/{id}:
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdate"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /activity:
    get:
      operationId: listActivity
      tags: [activity]
      summary: Get recent activity logs
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: candidateId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Activity logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityLog"

  /kanban:
    get:
      operationId: getKanban
      tags: [kanban]
      summary: Get all candidates grouped by status for Kanban view
      responses:
        "200":
          description: Kanban board data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KanbanBoard"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    CandidateStatus:
      type: string
      enum: [applied, screening, interview, offer, selected, rejected]

    Candidate:
      type: object
      required:
        - id
        - name
        - email
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        yearsOfExperience:
          type: ["integer", "null"]
        currentCompany:
          type: ["string", "null"]
        resumeUrl:
          type: ["string", "null"]
        position:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/CandidateStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CandidateDetail:
      type: object
      required:
        - id
        - name
        - email
        - status
        - createdAt
        - updatedAt
        - notes
        - activityLogs
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        yearsOfExperience:
          type: ["integer", "null"]
        currentCompany:
          type: ["string", "null"]
        resumeUrl:
          type: ["string", "null"]
        position:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/CandidateStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        notes:
          type: array
          items:
            $ref: "#/components/schemas/Note"
        activityLogs:
          type: array
          items:
            $ref: "#/components/schemas/ActivityLog"

    CandidateList:
      type: object
      required: [data, total, page, limit, totalPages]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Candidate"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer

    CandidateInput:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: string
        skills:
          type: array
          items:
            type: string
        yearsOfExperience:
          type: integer
          minimum: 0
        currentCompany:
          type: string
        resumeUrl:
          type: string
        position:
          type: string
        status:
          $ref: "#/components/schemas/CandidateStatus"

    CandidateUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: string
        skills:
          type: array
          items:
            type: string
        yearsOfExperience:
          type: integer
          minimum: 0
        currentCompany:
          type: string
        resumeUrl:
          type: string
        position:
          type: string
        status:
          $ref: "#/components/schemas/CandidateStatus"

    CandidateStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/CandidateStatus"

    Note:
      type: object
      required: [id, candidateId, content, authorName, createdAt, updatedAt]
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        content:
          type: string
        authorName:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    NoteInput:
      type: object
      required: [content, authorName]
      properties:
        content:
          type: string
          minLength: 1
        authorName:
          type: string
          minLength: 1

    NoteUpdate:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1

    ActivityLog:
      type: object
      required: [id, candidateId, candidateName, action, createdAt]
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        candidateName:
          type: string
        action:
          type: string
        oldValue:
          type: ["string", "null"]
        newValue:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    DashboardStats:
      type: object
      required:
        - totalCandidates
        - byStatus
        - recentCandidates
        - recentActivity
      properties:
        totalCandidates:
          type: integer
        byStatus:
          $ref: "#/components/schemas/StatusCounts"
        recentCandidates:
          type: array
          items:
            $ref: "#/components/schemas/Candidate"
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityLog"

    StatusCounts:
      type: object
      required: [applied, screening, interview, offer, selected, rejected]
      properties:
        applied:
          type: integer
        screening:
          type: integer
        interview:
          type: integer
        offer:
          type: integer
        selected:
          type: integer
        rejected:
          type: integer

    KanbanColumn:
      type: object
      required: [status, candidates]
      properties:
        status:
          $ref: "#/components/schemas/CandidateStatus"
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/Candidate"

    KanbanBoard:
      type: object
      required: [columns]
      properties:
        columns:
          type: array
          items:
            $ref: "#/components/schemas/KanbanColumn"