# lib/api-zod/src/generated/types/activityLog.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Enterprise Candidate Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityLog {
  id: number;
  candidateId: number;
  candidateName: string;
  action: string;
  /** @nullable */
  oldValue?: string | null;
  /** @nullable */
  newValue?: string | null;
  createdAt: Date;
}