# lib/api-zod/src/generated/types/candidateList.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Enterprise Candidate Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { Candidate } from './candidate';

export interface CandidateList {
  data: Candidate[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}