# lib/api-zod/src/generated/types/candidateStatus.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Enterprise Candidate Management System API
 * OpenAPI spec version: 0.1.0
 */

export type CandidateStatus = typeof CandidateStatus[keyof typeof CandidateStatus];


export const CandidateStatus = {
  applied: 'applied',
  screening: 'screening',
  interview: 'interview',
  offer: 'offer',
  selected: 'selected',
  rejected: 'rejected',
} as const;