# lib/api-zod/src/generated/types/candidateStatusUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Enterprise Candidate Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { CandidateStatus } from './candidateStatus';

export interface CandidateStatusUpdate {
  status: CandidateStatus;
}