# lib/api-zod/src/generated/types/noteUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Enterprise Candidate Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface NoteUpdate {
  /** @minLength 1 */
  content: string;
}